# v1.8.0: Complete Refactoring (Phases 1-4) - Command Pattern Architecture + Full Documentation