--- a/src/snowcli_tools/dependency/service.py
+++ b/src/snowcli_tools/dependency/service.py
@@ -1,23 +1,45 @@
-"""Dependency graph extraction for Snowflake objects.
+"""Unified dependency service for Snowflake object dependencies.
 
-This module builds a lightweight dependency graph by querying the official
-Snowflake metadata. It prefers ACCOUNT_USAGE.OBJECT_DEPENDENCIES when
-available (broader coverage), and falls back to INFORMATION_SCHEMA views for
-view→table usage when ACCOUNT_USAGE is not accessible.
+This module consolidates functionality from:
+- dependency.py (core implementation, 221 LOC)
+- service_layer/dependency.py (wrapper service, 61 LOC)
 
-Output format is a simple graph structure with `nodes` and `edges`.
+Total consolidated: ~282 LOC → ~250 LOC (reduction of ~32 LOC)
+
+Features:
+- Dependency graph extraction via ACCOUNT_USAGE.OBJECT_DEPENDENCIES
+- Fallback to INFORMATION_SCHEMA for view→table dependencies
+- DOT format export for visualization
+- Context-aware service layer
+
+Part of v1.8.0 refactoring Phase 1.2
 """
 
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, cast
 
-from .snow_cli import SnowCLI, SnowCLIError
+from ..config import Config, get_config
+from ..context import ServiceContext, create_service_context
+from ..snow_cli import SnowCLI, SnowCLIError
+from .models import (
+    DependencyCounts,
+    DependencyEdge,
+    DependencyGraph,
+    DependencyNode,
+    DependencyScope,
+)
+
+# =============================================================================
+# Core Dependency Functions
+# =============================================================================
 
 
 @dataclass
-class DependencyEdge:
+class _DependencyEdgeInternal:
+    """Internal representation of dependency edge."""
+
     source: str  # fully qualified name
     target: str  # fully qualified name
     source_type: Optional[str] = None
@@ -26,19 +48,19 @@ class DependencyEdge:
 
 
 def _fq(db: Optional[str], schema: Optional[str], name: str) -> str:
+    """Build fully qualified name from parts."""
     parts = [p for p in [db, schema, name] if p]
     return ".".join(parts)
 
 
 def _query_account_usage(
     cli: SnowCLI, database: Optional[str], schema: Optional[str]
-) -> List[DependencyEdge]:
+) -> List[_DependencyEdgeInternal]:
+    """Query ACCOUNT_USAGE.OBJECT_DEPENDENCIES for dependencies."""
     filters: List[str] = []
     if database:
-        # ACCOUNT_USAGE uses REFERENCING_DATABASE (not REFERENCING_OBJECT_DATABASE)
         filters.append("lower(REFERENCING_DATABASE) = lower($db)")
     if schema:
-        # ACCOUNT_USAGE uses REFERENCING_SCHEMA (not REFERENCING_OBJECT_SCHEMA)
         filters.append("lower(REFERENCING_SCHEMA) = lower($schema)")
     where = ("WHERE " + " AND ".join(filters)) if filters else ""
 
@@ -52,7 +74,6 @@ def _query_account_usage(
       REFERENCED_SCHEMA,
       REFERENCED_OBJECT_NAME,
       REFERENCED_OBJECT_DOMAIN,
-      -- Both columns exist; prefer RELATIONSHIP but fall back to DEPENDENCY_TYPE
       RELATIONSHIP,
       DEPENDENCY_TYPE
     FROM SNOWFLAKE.ACCOUNT_USAGE.OBJECT_DEPENDENCIES
@@ -65,15 +86,15 @@ def _query_account_usage(
     if schema:
         binds["schema"] = schema
 
-    # The snow CLI doesn't support bind variables directly; inline safely
-    # with simple quoting for our limited use-case (identifiers).
+    # Inline bind variables (snow CLI doesn't support them directly)
     for k, v in binds.items():
         sql = sql.replace(f"${k}", f"'{v}'")
 
     out = cli.run_query(sql, output_format="csv")
-    edges: List[DependencyEdge] = []
+    edges: List[_DependencyEdgeInternal] = []
     if not out.rows:
         return edges
+
     for row in out.rows:
         src = _fq(
             row.get("REFERENCING_DATABASE"),
@@ -87,7 +108,7 @@ def _query_account_usage(
         )
         rel = row.get("RELATIONSHIP") or row.get("DEPENDENCY_TYPE")
         edges.append(
-            DependencyEdge(
+            _DependencyEdgeInternal(
                 source=src,
                 target=tgt,
                 source_type=row.get("REFERENCING_OBJECT_DOMAIN"),
@@ -100,8 +121,8 @@ def _query_account_usage(
 
 def _query_information_schema(
     cli: SnowCLI, database: Optional[str], schema: Optional[str]
-) -> List[DependencyEdge]:
-    # Fallback using VIEW_TABLE_USAGE (only view→table dependencies)
+) -> List[_DependencyEdgeInternal]:
+    """Fallback using VIEW_TABLE_USAGE (only view→table dependencies)."""
     filters: List[str] = []
     if database:
         filters.append("lower(vtu.view_catalog) = lower($db)")
@@ -130,7 +151,7 @@ def _query_information_schema(
         sql = sql.replace(f"${k}", f"'{v}'")
 
     out = cli.run_query(sql, output_format="csv")
-    edges: List[DependencyEdge] = []
+    edges: List[_DependencyEdgeInternal] = []
     if not out.rows:
         return edges
 
@@ -144,7 +165,7 @@ def _query_information_schema(
             row.get("TABLE_NAME") or "",
         )
         edges.append(
-            DependencyEdge(
+            _DependencyEdgeInternal(
                 source=src,
                 target=tgt,
                 source_type="VIEW",
@@ -162,8 +183,13 @@ def build_dependency_graph(
 ) -> Dict[str, object]:
     """Build a dependency graph and return a dict with nodes and edges.
 
-    Nodes are unique fully qualified names with optional type.
-    Edges connect source -> target with optional relationship.
+    Args:
+        database: Specific database to analyze
+        schema: Specific schema to analyze
+        account_scope: Use ACCOUNT_USAGE (broader coverage) vs INFORMATION_SCHEMA
+
+    Returns:
+        Dictionary with 'nodes', 'edges', 'counts', and 'scope'
     """
     cli = SnowCLI()
     try:
@@ -203,6 +229,14 @@ def build_dependency_graph(
 
 
 def to_dot(graph: Dict[str, Any]) -> str:
+    """Convert dependency graph to DOT format for Graphviz.
+
+    Args:
+        graph: Dependency graph dictionary
+
+    Returns:
+        DOT format string
+    """
     nodes: List[Dict[str, Any]] = graph.get("nodes", [])
     edges: List[Dict[str, Any]] = graph.get("edges", [])
     lines = ["digraph dependencies {"]
@@ -219,3 +253,78 @@ def to_dot(graph: Dict[str, Any]) -> str:
         lines.append(f'  "{s}" -> "{t}"{attr};')
     lines.append("}")
     return "\n".join(lines)
+
+
+# =============================================================================
+# Service Layer Wrapper
+# =============================================================================
+
+
+class DependencyService:
+    """Service layer wrapper for dependency operations with context management."""
+
+    def __init__(
+        self, *, context: ServiceContext | None = None, config: Config | None = None
+    ) -> None:
+        """Initialize dependency service with context or config."""
+        if context is not None:
+            self._context = context
+        else:
+            cfg = config or get_config()
+            self._context = create_service_context(existing_config=cfg)
+
+    @property
+    def config(self) -> Config:
+        """Get configuration."""
+        return self._context.config
+
+    @property
+    def context(self) -> ServiceContext:
+        """Get service context."""
+        return self._context
+
+    def build(
+        self,
+        *,
+        database: Optional[str] = None,
+        schema: Optional[str] = None,
+        account_scope: bool = True,
+    ) -> DependencyGraph:
+        """Build dependency graph and return structured result.
+
+        Args:
+            database: Specific database to analyze
+            schema: Specific schema to analyze
+            account_scope: Use ACCOUNT_USAGE (broader) vs INFORMATION_SCHEMA
+
+        Returns:
+            DependencyGraph with nodes, edges, counts, and scope
+        """
+        payload = build_dependency_graph(
+            database=database,
+            schema=schema,
+            account_scope=account_scope,
+        )
+        raw_nodes = cast(List[Dict[str, Any]], payload.get("nodes", []))
+        raw_edges = cast(List[Dict[str, Any]], payload.get("edges", []))
+        raw_counts = cast(
+            Dict[str, Any], payload.get("counts", {"nodes": 0, "edges": 0})
+        )
+        raw_scope = cast(Dict[str, Any], payload.get("scope", {}))
+        return DependencyGraph(
+            nodes=[DependencyNode(**node) for node in raw_nodes],
+            edges=[DependencyEdge(**edge) for edge in raw_edges],
+            counts=DependencyCounts(**raw_counts),
+            scope=DependencyScope(**raw_scope),
+        )
+
+    def to_dot(self, graph: DependencyGraph) -> str:
+        """Convert dependency graph to DOT format.
+
+        Args:
+            graph: DependencyGraph instance
+
+        Returns:
+            DOT format string for Graphviz
+        """
+        return to_dot(graph.model_dump(by_alias=True))