--- a/src/snowcli_tools/mcp/tools/__init__.py
+++ b/src/snowcli_tools/mcp/tools/__init__.py
@@ -12,23 +12,30 @@ Part of v1.8.0 Phase 2.2 - MCP Server Simplification
 from __future__ import annotations
 
 from .base import MCPTool, MCPToolSchema
+from .build_catalog import BuildCatalogTool
+from .build_dependency_graph import BuildDependencyGraphTool
+from .check_profile_config import CheckProfileConfigTool
+from .check_resource_dependencies import CheckResourceDependenciesTool
 from .execute_query import ExecuteQueryTool
+from .get_catalog_summary import GetCatalogSummaryTool
+from .get_resource_status import GetResourceStatusTool
 from .health_check import HealthCheckTool
-
-# TODO: Extract remaining tools:
-# - PreviewTableTool
-# - BuildCatalogTool
-# - QueryLineageTool
-# - BuildDependencyGraphTool
-# - TestConnectionTool
-# - GetCatalogSummaryTool
-# - CheckProfileConfigTool
-# - GetResourceStatusTool
-# - CheckResourceDependenciesTool
+from .preview_table import PreviewTableTool
+from .query_lineage import QueryLineageTool
+from .test_connection import TestConnectionTool
 
 __all__ = [
     "MCPTool",
     "MCPToolSchema",
+    "BuildCatalogTool",
+    "BuildDependencyGraphTool",
+    "CheckProfileConfigTool",
+    "CheckResourceDependenciesTool",
     "ExecuteQueryTool",
+    "GetCatalogSummaryTool",
+    "GetResourceStatusTool",
     "HealthCheckTool",
+    "PreviewTableTool",
+    "QueryLineageTool",
+    "TestConnectionTool",
 ]