… PHASE2.3_COMPLETE.md
… PHASE2_COMPLETE.md
… PHASE3_PROGRESS.md
… PHASE4_COMPLETE.md
… REFACTORING_SUMMARY.md
… V1.8.0_FINAL_SUMMARY.md
… docs/agentic_workflows/README.md
… docs/agentic_workflows/agent_configurations.md
… docs/agentic_workflows/prompt_engineering_guide.md
… docs/api/ERROR_CATALOG.md
… docs/api/README.md
… docs/api/TOOLS_INDEX.md
… docs/api/tools/build_catalog.md
… docs/api/tools/execute_query.md
… docs/api/tools/health_check.md
… docs/api/tools/preview_table.md
… docs/api/tools/test_connection.md
… pyproject.toml
… src/snowcli_tools/catalog.py
… src/snowcli_tools/catalog/service.py
… src/snowcli_tools/catalog_service.py
… src/snowcli_tools/dependency/models.py
… src/snowcli_tools/mcp/tools/check_profile_config.py
… src/snowcli_tools/mcp/tools/check_resource_dependencies.py
… src/snowcli_tools/mcp/tools/execute_query.py
… src/snowcli_tools/mcp/tools/get_catalog_summary.py
… src/snowcli_tools/mcp/tools/get_resource_status.py
… src/snowcli_tools/mcp/tools/health_check.py
… src/snowcli_tools/mcp/tools/preview_table.py
… src/snowcli_tools/mcp/tools/query_lineage.py